cfxkey

Struct Secret

pub struct Secret { /* private fields */ }

Implementations

impl Secret

pub fn from_slice(key: &[u8]) -> Option<Self>

Creates a Secret from the given slice, returning None if the slice length != 32.

pub fn zero() -> Self

Creates zero key, which is invalid for crypto operations, but valid for math operation.

pub fn from_unsafe_slice(key: &[u8]) -> Result<Self, Error>

Imports and validates the key.

pub fn check_validity(&self) -> Result<(), Error>

Checks validity of this key.

pub fn add(&mut self, other: &Secret) -> Result<(), Error>

Inplace add one secret key to another (scalar + scalar)

pub fn sub(&mut self, other: &Secret) -> Result<(), Error>

Inplace subtract one secret key from another (scalar - scalar)

pub fn dec(&mut self) -> Result<(), Error>

Inplace decrease secret key (scalar - 1)

pub fn mul(&mut self, other: &Secret) -> Result<(), Error>

Inplace multiply one secret key to another (scalar * scalar)

pub fn neg(&mut self) -> Result<(), Error>

Inplace negate secret key (-scalar)

pub fn inv(&mut self) -> Result<(), Error>

Inplace inverse secret key (1 / scalar)

pub fn pow(&mut self, pow: usize) -> Result<(), Error>

Compute power of secret key inplace (secret ^ pow). This function is not intended to be used with large powers.

pub fn to_secp256k1_secret(&self) -> Result<SecretKey, Error>

Create secp256k1::key::SecretKey based on this secret

pub fn to_hex(&self) -> String